[issue #36] remove thread to runnable assoc
The runnable was associated to the actual thread running it in
beforeExecute. However, we don't need this association because the
executor actually uses the currently running thread to execute the task
that is eventually finished when afterExecute is called. Thus the "where
is the thread" issue is not present anymore.