[issue #43] remove oracle dbobject hack, dbstring adaption
The dbobject hack for oracle is removed due to the changes in the server
persistence manager. Effectively, the dbobject operation is currently
unused. The dbstring is adapted so that it only returns the wkt of the
geometry because it is not possible to provide oracle driver with a
proper db string. It now relies on the fact that a proper custom
insert/update statement is available for the table that contains the
geometry. c3p0 is not needed anymore due to the remove of the hack.